1. Upgraded the dependencies 2. Changed the folder structure
