+++ .prettierrc
@@ -1,5 +1,5 @@
 {
     "singleQuote": true,
     "printWidth": 100,
-    "tabWidth": 4
+    "tabWidth": 2
 }
