+++ package.json
@@ -89,55 +89,57 @@
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
-    "@hot-loader/react-dom": "^16.8.6",
+    "@hot-loader/react-dom": "^16.9.0",
     "@types/electron-devtools-installer": "^2.2.0",
-    "@types/jest": "^24.0.13",
-    "@types/react": "^16.8.18",
-    "@types/react-dom": "^16.8.4",
+    "@types/jest": "^24.0.18",
+    "@types/react": "^16.9.2",
+    "@types/react-dom": "^16.9.0",
     "@types/react-hot-loader": "^4.1.0",
-    "@types/react-redux": "^7.0.9",
-    "@types/react-test-renderer": "^16.8.1",
-    "@types/webdriverio": "^4.8.7",
+    "@types/react-redux": "^7.1.2",
+    "@types/react-test-renderer": "^16.9.0",
+    "@types/webdriverio": "^5.0.0",
     "@types/webpack-env": "^1.13.3",
     "babel-loader": "^8.0.6",
-    "cross-env": "^5.1.3",
-    "css-loader": "^2.1.1",
+    "cross-env": "^5.2.1",
+    "css-loader": "^3.2.0",
     "electron": "^5.0.8",
-    "electron-builder": "^21.1.5",
+    "electron-builder": "^21.2.0",
     "electron-devtools-installer": "^2.2.4",
-    "file-loader": "^3.0.1",
-    "fork-ts-checker-webpack-plugin": "^1.3.4",
+    "file-loader": "^4.2.0",
+    "fork-ts-checker-webpack-plugin": "^1.5.0",
     "html-webpack-plugin": "^3.2.0",
-    "husky": "^2.3.0",
-    "image-webpack-loader": "^4.6.0",
-    "jest": "^24.8.0",
-    "lint-staged": "^8.1.7",
+    "husky": "^3.0.5",
+    "image-webpack-loader": "^5.0.0",
+    "jest": "^24.9.0",
+    "lint-staged": "^9.2.5",
     "node-sass": "^4.12.0",
     "prettier": "^1.17.1",
-    "react-hot-loader": "^4.8.8",
-    "react-test-renderer": "^16.8.6",
+    "react-hot-loader": "^4.12.12",
+    "react-test-renderer": "^16.9.0",
     "redux-devtools-extension": "^2.13.5",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^8.0.0",
     "source-map-loader": "^0.2.4",
-    "spectron": "^5.0.0",
-    "style-loader": "^0.23.1",
+    "spectron": "^8.0.0",
+    "style-loader": "^1.0.0",
     "ts-jest": "^24.0.2",
-    "tslint": "^5.16.0",
+    "tslint": "^5.19.0",
     "tslint-config-airbnb": "^5.11.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-react": "^4.0.0",
-    "typescript": "^3.4.5",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.4.1",
-    "webpack-merge": "^4.2.1"
+    "typescript": "^3.6.2",
+    "webpack": "^4.39.3",
+    "webpack-cli": "^3.3.7",
+    "webpack-dev-server": "^3.8.0",
+    "webpack-merge": "^4.2.2"
   },
   "dependencies": {
     "electron-store": "^4.0.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-redux": "^7.0.3",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-redux": "^7.1.1",
     "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0",
+    "reselect": "^4.0.0",
     "wine": "^0.9.8"
   }
 }
