+++ src/components/Counter.scss
@@ -1,0 +1,3 @@
+.counter {
+  background: white;
+}
