+++ src/reduxContent/wallet/types.ts
@@ -1,0 +1,3 @@
+export const SETTING = 'SETTING';
+export const INCREMENT = 'INCREMENT';
+export const DECREMENT = 'DECREMENT';
