--- src/renderer/actions/index.ts
@@ -1,3 +1,0 @@
-import { CounterAction } from './counterActions';
-
-export type RootActions = CounterAction[keyof CounterAction];
