--- src/renderer/components/Counter.scss
@@ -1,3 +1,0 @@
-.counter {
-    background: white;
-}
