+++ test/e2e/example.spec.ts
@@ -4,52 +4,52 @@ import * as path from 'path';
 
 jest.setTimeout(10000);
 
-describe('Main window', () => {
-    let app: Application;
+// describe('Main window', () => {
+//   let app: Application;
 
-    beforeEach(() => {
-        app = new Application({
-            path: electronPath.toString(),
-            args: [path.join(__dirname, '..', '..')]
-        });
+//   beforeEach(() => {
+//     app = new Application({
+//       path: electronPath.toString(),
+//       args: [path.join(__dirname, '..', '..')]
+//     });
 
-        return app.start();
-    });
+//     return app.start();
+//   });
 
-    afterEach(() => {
-        if (app.isRunning()) {
-            return app.stop();
-        }
-    });
+//   afterEach(() => {
+//     if (app.isRunning()) {
+//       return app.stop();
+//     }
+//   });
 
-    it('opens the window', async () => {
-        const { client, browserWindow } = app;
+//   it('opens the window', async () => {
+//     const { client, browserWindow } = app;
 
-        await client.waitUntilWindowLoaded();
-        const title = await browserWindow.getTitle();
+//     await client.waitUntilWindowLoaded();
+//     const title = await browserWindow.getTitle();
 
-        expect(title).toBe('Cryptonomic');
-    });
+//     expect(title).toBe('Cryptonomic');
+//   });
 
-    it('increments the counter', async () => {
-        const { client } = app;
+//   it('increments the counter', async () => {
+//     const { client } = app;
 
-        await client.waitUntilWindowLoaded();
-        await client.click('#increment');
+//     await client.waitUntilWindowLoaded();
+//     await client.click('#increment');
 
-        const counterText = await client.getText('#counter-value');
+//     const counterText = await client.getText('#counter-value');
 
-        expect(counterText).toBe('Current value: 1');
-    });
+//     expect(counterText).toBe('Current value: 1');
+//   });
 
-    it('decrements the counter', async () => {
-        const { client } = app;
+//   it('decrements the counter', async () => {
+//     const { client } = app;
 
-        await client.waitUntilWindowLoaded();
-        await client.click('#decrement');
+//     await client.waitUntilWindowLoaded();
+//     await client.click('#decrement');
 
-        const counterText = await client.getText('#counter-value');
+//     const counterText = await client.getText('#counter-value');
 
-        expect(counterText).toBe('Current value: -1');
-    });
-});
+//     expect(counterText).toBe('Current value: -1');
+//   });
+// });
